Decouple Ant task for protobuf from the main Ant task
With this CL, the Ant task for protobuf is separately executed from the main Ant task to build Mozc APK, which allows us to improve the build concurrency and makes android.gyp a bit simpler.

This is just a clean-up of build rules.  Hence no behavior change is intended.

Patch from Hiroyuki Komatsu.

BUG=none
TEST=compile


git-svn-id: https://mozc.googlecode.com/svn/trunk@416 a6090854-d499-a067-5803-1114d4e51264