Explicitly enable PIE in standaline executables for unittest on Android
Android 5.0 requires standalone native executable to be PIE, which means that we need to build unittest executables to with PIE enabled.  Note that this change has nothing to do with the production, where native code is used only as a shared library.

This CL also disables native unittest on Android if API level is 14 or 15, which do not support PIE.  Probably we could work around that limitation by using the same approach in Chromium, but currently we simply disable unittest for those platforms.

Reference:
- https://code.google.com/p/android-developer-preview/issues/detail?id=888
- https://code.google.com/p/chromium/issues/detail?id=373219

Patch from Tsuyoshi Matsuzaki.

BUG=none
TEST=compile


git-svn-id: https://mozc.googlecode.com/svn/trunk@427 a6090854-d499-a067-5803-1114d4e51264