Suppress usage stats dialog in test environment for better performance and stability
Usage stats dialog in test environment will be disabled with this CL.

Usually the dialog in question will not be closed by anybody if it is shown in invisible X session.  In this case emulator process will never die and it keeps using provided port.  This will block invocation of emulator because of no available port.

This CL has nothing to do with the production build.  Hence no behavior change in production is intended.

Patch from Tsuyoshi Matsuzaki.

BUG=none
TEST=none


git-svn-id: https://mozc.googlecode.com/svn/trunk@430 a6090854-d499-a067-5803-1114d4e51264