Fix resource leak on onDestroy
Starting from Android 5.0, IME processes can be alive even after Service.onDestory is called depending on the available memory so that we can switch IMEs as fast as possible.
https://android.googlesource.com/platform/frameworks/base/+/f0f94d129b6eb3c48624e915898d86d4f2de59ff

However, this change revealed that Mozc has not released all the Java objects on Service.onDestory, which can be observed as an increasing memory usage of Mozc.

Patch from Tsuyoshi Matsuzaki.

Closes Issue 265.

BUG=Issue mozc:265 
TEST=manually done with Nexus 5 / Android 5.0.1 (LRX22C)


git-svn-id: https://mozc.googlecode.com/svn/trunk@474 a6090854-d499-a067-5803-1114d4e51264