Remove a workaround for sysconf w/ _SC_PHYS_PAGES
sysconf with _SC_PHYS_PAGES did not work on Android 2.2 and prior due to a bug in bionic. That bug in question was fixed in Android 2.3 as follows.
https://android.googlesource.com/platform/bionic/+/038fbae518e904c7aba64779714a22dbeeb90887

Mozc has worked around this issue by emulating sysconf with _SC_PHYS_PAGES in android/jni/sysconf.cc.  However, we no longer have to do that because Mozc don’t support those legacy platforms anymore.


This CL is just a removal of redundant historical code.  No behavior change is intended.

BUG=none
TEST=none


git-svn-id: https://mozc.googlecode.com/svn/trunk@540 a6090854-d499-a067-5803-1114d4e51264