Do not specify full-width annotation for Emojis
With this CL, Util::GetScriptType() categorizes Google PUA Emojis
as Emojis in all the platforms so that variants rewriter does not
specify full-width annotation for Android.  Such an assumption
against Unicode PUA would be suboptimal for other platforms though.

This CL also updates EmoticonRewriter so as not to specify
character-width anotation for emoticon candidates.

BUG=none
TEST=unittest